--- polls/templates/registration/logged_out.html
@@ -1,7 +1,0 @@
-   <html>
-   <body>
-   <h2>Logged out</h2>
-   <p>
-   Want something todo?
-   <a href="{% url 'login'%}">Login</a> again
-   </body>
