--- polls/templates/registration/login.html
@@ -1,17 +1,0 @@
-<html>
- <body>
- <h2>Poll: Login</h2>
- <form method="post">
-   {% csrf_token %}
-   {{ form.as_p }}
-   <p>
-   <button type="submit">Login</button>
-   </p>
-   <input type="hidden" name="next" value="{{next}}"/>
- </form>
- {# If you have written a password_reset template, then add #}
- <p>
- <a href="{% url 'password_reset' %}">Forgot password?</a>
- </p>
- </body>
- </html>
