--- polls/test/test_index.py
@@ -1,67 +1,0 @@
-"""Module for test the index page."""
-
-import datetime
-
-from django.test import TestCase
-from django.utils import timezone
-from django.urls import reverse
-
-from polls.models import Question
-
-
-def create_question(question_text, days):
-    """
-    Create a question with the given `question_text` and published date.
-
-    The given number of `days` offset to now (negative for questions published
-    in the past, positive for questions that have yet to be published).
-    """
-    time = timezone.now() + datetime.timedelta(days=days)
-    end_time = timezone.now() + datetime.timedelta(days=30)
-    return Question.objects.create(question_text=question_text, pub_date=time, end_date=end_time)
-
-
-class QuestionIndexViewTests(TestCase):
-    """Test the Question index page."""
-
-    def test_no_questions(self):
-        """If no questions exist, an appropriate message is displayed."""
-        response = self.client.get(reverse('polls:index'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_past_question(self):
-        """pub_date in the past question are displayed on the index page."""
-        create_question(question_text="Past question.", days=-30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'], ['<Question: Past question.>']
-        )
-
-    def test_future_question(self):
-        """pub_date in the future question not displayed on the index page."""
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_future_question_and_past_question(self):
-        """Only past questions are displayed."""
-        create_question(question_text="Past question.", days=-30)
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_two_past_questions(self):
-        """The questions index page may display multiple questions."""
-        create_question(question_text="Past question 1.", days=-30)
-        create_question(question_text="Past question 2.", days=-5)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question 2.>', '<Question: Past question 1.>']
-        )
